+++ src/components/game/graphics/Pipeline.js
@@ -1,5 +1,11 @@
 // Graphics building blocks to render the game
 
+// Every action in graphics that draws shapes will boil down to a sequence
+// of function calls from the functions below. As of now they deal with JS/HTML 
+// Canvas, which is very limited and slow, but in the future these functions will 
+// serve as entry points to call Rust/WASM/WebGL functions and APIs for more
+// powerful rendering.
+
 export function drawRect(canvas, x, y, xSize, ySize, color) {
   canvas.fillStyle = color;
   canvas.fillRect(x, y, xSize, ySize);
