+++ src/components/game/graphics/objects/stage/Stage.js
@@ -1,6 +1,6 @@
-import { STAGE_WIDTH, WINDOW_SIZE } from "../../Constants";
-import { drawCell } from "../objects/cell/DrawCell";
-import { drawRect, outlineRect } from "../utils/Shapes";
+import { STAGE_WIDTH, WINDOW_SIZE } from "../../../Constants";
+import { drawCell } from "../cell/DrawCell";
+import { drawRect, outlineRect } from "../../Pipeline";
 
 const STAGE_CELL_SIZE = WINDOW_SIZE / 32;
